bot/requirements.txt
requests~=2.28.2
python-dotenv==0.19.0
aiogram~=2.25.1
aiopg~=1.4.0
beautifulsoup4~=4.11.2